runpod-service/requirements.txt
runpod
transformers
torch
numpy
python-dotenv
grpcio
grpcio-tools
protobuf
grpcio-reflection
accelerate
bitsandbytes
sentencepiece
pillow
